backend/app/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
boto3==1.34.34
botocore==1.34.34
Django==4.2
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
